Класс, описывающий боевые характеристики врагов
