+++ Assembly-CSharp.csproj
@@ -85,7 +85,6 @@
     <Compile Include="Assets\AlexRK2\BackHandler.cs" />
     <Compile Include="Assets\AlexRK2\RKDialogueObserver.cs" />
     <Compile Include="Assets\BattleSystem\BattleStats.cs" />
-    <Compile Include="Assets\BattleSystem\PlayerBattleStats.cs" />
     <Compile Include="Assets\BattleSystem\Solidity.cs" />
     <Compile Include="Assets\BattleSystem\WeaponInfo.cs" />
     <Compile Include="Assets\CameraClickHandler.cs" />
